refactor(schemes): export scheme options directly in scheme file
