+++ src/providers/auth0/contracts/Auth0ProviderOptions.ts
@@ -1,4 +1,4 @@
-import Oauth2SchemeOptions from '../../../schemes/oauth2/contracts/Oauth2SchemeOptions'
+import { Oauth2SchemeOptions } from '../../../schemes/oauth2'
 import { ProviderOptions } from '../../index'
 
 export interface Auth0ProviderOptions
