+++ src/schemes/cookie/contracts/CookieSchemeOptions.ts
@@ -1,7 +1,4 @@
-import {
-  LocalSchemeEndpoints,
-  LocalSchemeOptions
-} from '../../local/contracts/LocalSchemeOptions'
+import { LocalSchemeEndpoints, LocalSchemeOptions } from '../../local'
 import { HTTPRequest } from '../../../index'
 
 export interface CookieSchemeEndpoints extends LocalSchemeEndpoints {
